pc_server/requirements.txt
fastapi
uvicorn[standard]
python-socketio
psutil
